% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliR-package.R
\docType{package}
\name{sliR-package}
\alias{sliR}
\alias{sliR-package}
\title{sliR: Standardized Length Index and Allometric Data Simulation}
\description{
Computes the Standardized Length Index (SLI) for size-correcting appendage lengths, following the scaled mass index approach of Peig and Green (2009) \doi{10.1111/j.1600-0706.2009.17643.x}, and simulates allometric morphological data with controllable scaling exponents, correlation structure, and measurement or transient error.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LezzGitIt/sliR}
  \item Report bugs at \url{https://github.com/LezzGitIt/sliR/issues}
}

}
\author{
\strong{Maintainer}: Aaron Skinner \email{skinnerayayron93@gmail.com}

}
\keyword{internal}
